+++ src/main/java/com/kswr/wallets/api/walletsapi/domain/Wallet.java
@@ -21,7 +21,7 @@ public class Wallet {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @NotBlank
     private String name;
@@ -32,7 +32,7 @@ public class Wallet {
     private Geometry geo;
 
     @Builder
-    public Wallet(Integer id, @NotBlank String name, Geometry geo) {
+    public Wallet(Long id, @NotBlank String name, Geometry geo) {
         this.id = id;
         this.name = name;
         this.geo = geo;
